+++ RELEASE_NOTES.md
@@ -4,6 +4,8 @@ Changes:
 
   * Added FTT_152C1_1WH slide label layout 
   * Added RIN and DV200 for samples in Pinery-MISO
+  * Added interface for administrators to add, edit, and delete sample classes and relationships
+    (detailed sample)
 
 # 0.2.194